+++ src/inet/BUILD.gn
@@ -84,7 +84,6 @@ static_library("inet") {
     "InetInterface.h",
     "InetLayer.cpp",
     "InetLayer.h",
-    "InetLayerBasis.h",
     "InetLayerEvents.h",
     "arpa-inet-compatibility.h",
   ]
